Last header not being copied in parser.py for GET
In method create_request of class RequestCreator in file
syntribos/clients/http/parser.py line 43, the headers variable
is missing the last HTTP header when a GET file is being processed.
The code works for POST because the POST body is separated from the
header by two newlines. For GET requests, the index value is being
adjusted to account for the fact that it does not have a body so no
empty line will be present.  We just add 1 to the index in that case.

Change-Id: I91881e1882aab21c2c14a7ff351bb3f7b70727f2
Closes-Bug: #1543236