Adding request/response content compression
Adding request/response content compression and
modifying the print_log method to display the update log_dir

This is part of the change to update th overall structure of the
debug log. Adding a logger options to config.

Added compression to request params and headers as well, this has
helped in reducing avg log size to abt 50 percent, without loss of
information.

Updated config file to reflect this change
Change-Id: Ifac7a671607945a4ca15d2e87b28b4b2718f0ee1