Removing overlapping payload/failure key
One of the payloads in the LDAP fuzz strings can trigger false
positives with the command injection test, due to an overlap between
the payload and the failure keys of the command injection test.

Tried introducing a unittest to prevent this, but utilizing payloads
files in unittests is too tricky at this point.

Change-Id: I8ea1782d008438876fbd77dbacfdffb2f3b70ae2