Glance Images Templates
Adds templates for Glance images + members. Also tweaks the HTTP parser
so that it doesn't break with a PATCH body consisting of a list instead
of the more-common dict. This solution is sub-optimal because the body
is treated as a string variable (i.e. you'll get a fuzzed request like
[0x0000] instead of fuzzing the contained object attributes)

Change-Id: I0e23c8281a63018ecadb84959f3a70be1f80fa74