Fixing up SSL test
The SSL test was not preparing the test request before sending it off
to the target server. This resulted in requests being sent in the wrong
encoding format, or with "{}" around URL variables.

This patch removes the test request, since it is the same as the
initial request, and instead tests the initial request for insecure
links.

Change-Id: Ib7344719dd84c83411600628e43939e236570160